requirements.txt
django>=4.2,<5.0
djangorestframework>=3.14
django-cors-headers>=4.0
Pillow>=10.0
numpy>=1.24
opencv-python-headless>=4.8
scikit-image>=0.21
tensorflow>=2.13
keras>=2.13
ultralytics>=8.0
twilio>=8.0
gunicorn>=21.0
